Modify error message so that it doesn't cause vs-tac build to fail
vs-tac uses regex to match strings in the cordova build output and treat them as fatal errors. In the case of being unable to create the universal links file, which cannot be created with a wildcard on windows machines, we don't want it to cause the build to fail. Change the error message so that it won't be detected by vs-tac, which checks for:
NO_FILE_DETECT = "Error: ENOENT, no such file or directory";